pyproject.toml
[tool.poetry]
name = "dz-12-2(5-6)"
version = "0.1.0"
description = ""
authors = ["Dmitry Krasheninnikov <ice_eyes@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
